refactor: retire layer packages
Horizontal packages made ownership depend on naming rather than
imports. Each presentation type now sits with the feature that
defines its behavior.
