--- mpvqc/dialogs/__init__.py
@@ -1,7 +1,0 @@
-# SPDX-FileCopyrightText: mpvQC developers
-#
-# SPDX-License-Identifier: GPL-3.0-or-later
-
-from .about import MpvqcAboutDialogViewModel as MpvqcAboutDialogViewModel
-from .edit_input import MpvqcEditInputDialogViewModel as MpvqcEditInputDialogViewModel
-from .edit_mpv import MpvqcEditMpvDialogViewModel as MpvqcEditMpvDialogViewModel
