+++ mpvqc/services/__init__.py
@@ -11,4 +11,10 @@ from .label_width_calculator import LabelWidthCalculatorService as LabelWidthCal
 from .resource import ResourceService as ResourceService
 from .settings_file import SettingsFileService as SettingsFileService
 from .state import StateService as StateService
+from .version_checker import HOME_URL as HOME_URL
+from .version_checker import CheckOutcome as CheckOutcome
+from .version_checker import NewVersionAvailable as NewVersionAvailable
+from .version_checker import ServerError as ServerError
+from .version_checker import ServerNotReachable as ServerNotReachable
+from .version_checker import UpToDate as UpToDate
 from .version_checker import VersionCheckerService as VersionCheckerService
