+++ mpvqc/shell/models/dependencies.py
