+++ mpvqc/shell/models/shortcuts.py
