+++ mpvqc/shell/viewmodels/quit.py
