+++ mpvqc/shell/viewmodels/reset.py
