--- mpvqc/viewmodels/__init__.py
@@ -1,13 +1,0 @@
-# SPDX-FileCopyrightText: mpvQC developers
-#
-# SPDX-License-Identifier: GPL-3.0-or-later
-
-from .message_boxes import MpvqcMessageBoxRequestRelayViewModel as MpvqcMessageBoxRequestRelayViewModel
-from .message_boxes import MpvqcQuitMessageBoxViewModel as MpvqcQuitMessageBoxViewModel
-from .message_boxes import MpvqcResetMessageBoxViewModel as MpvqcResetMessageBoxViewModel
-from .message_boxes import MpvqcVersionCheckMessageBoxViewModel as MpvqcVersionCheckMessageBoxViewModel
-from .utility import MpvqcFontsViewModel as MpvqcFontsViewModel
-from .views import MpvqcFooterViewModel as MpvqcFooterViewModel
-from .views import MpvqcHeaderViewModel as MpvqcHeaderViewModel
-from .views import MpvqcMenuBarViewModel as MpvqcMenuBarViewModel
-from .views import MpvqcToolBarViewModel as MpvqcToolBarViewModel
