--- mpvqc/viewmodels/message_boxes/__init__.py
@@ -1,8 +1,0 @@
-# SPDX-FileCopyrightText: mpvQC developers
-#
-# SPDX-License-Identifier: GPL-3.0-or-later
-
-from .quit import MpvqcQuitMessageBoxViewModel as MpvqcQuitMessageBoxViewModel
-from .request_relay import MpvqcMessageBoxRequestRelayViewModel as MpvqcMessageBoxRequestRelayViewModel
-from .reset import MpvqcResetMessageBoxViewModel as MpvqcResetMessageBoxViewModel
-from .version_check import MpvqcVersionCheckMessageBoxViewModel as MpvqcVersionCheckMessageBoxViewModel
