--- mpvqc/viewmodels/message_boxes/request_relay.py
@@ -1,27 +1,0 @@
-# SPDX-FileCopyrightText: mpvQC developers
-#
-# SPDX-License-Identifier: GPL-3.0-or-later
-
-import inject
-from PySide6.QtCore import QObject, Signal
-from PySide6.QtQml import QmlElement
-
-from mpvqc.exporting.services import ExportService
-from mpvqc.shell.services import QuitService
-
-QML_IMPORT_NAME = "io.github.mpvqc.mpvQC.Python"
-QML_IMPORT_MAJOR_VERSION = 1
-
-
-@QmlElement
-class MpvqcMessageBoxRequestRelayViewModel(QObject):
-    _exporter = inject.attr(ExportService)
-    _quit = inject.attr(QuitService)
-
-    exportErrorOccurred = Signal(str, int)
-    quitConfirmationNeeded = Signal()
-
-    def __init__(self, parent: QObject | None = None) -> None:
-        super().__init__(parent)
-        self._exporter.export_error_occurred.connect(self.exportErrorOccurred)
-        self._quit.confirmation_needed.connect(self.quitConfirmationNeeded)
