--- mpvqc/viewmodels/utility/__init__.py
@@ -1,5 +1,0 @@
-# SPDX-FileCopyrightText: mpvQC developers
-#
-# SPDX-License-Identifier: GPL-3.0-or-later
-
-from .fonts import MpvqcFontsViewModel as MpvqcFontsViewModel
