--- mpvqc/viewmodels/views/__init__.py
@@ -1,8 +1,0 @@
-# SPDX-FileCopyrightText: mpvQC developers
-#
-# SPDX-License-Identifier: GPL-3.0-or-later
-
-from .footer import MpvqcFooterViewModel as MpvqcFooterViewModel
-from .header import MpvqcHeaderViewModel as MpvqcHeaderViewModel
-from .header import MpvqcMenuBarViewModel as MpvqcMenuBarViewModel
-from .header import MpvqcToolBarViewModel as MpvqcToolBarViewModel
