--- mpvqc/viewmodels/views/header/__init__.py
@@ -1,8 +1,0 @@
-# SPDX-FileCopyrightText: mpvQC developers
-#
-# SPDX-License-Identifier: GPL-3.0-or-later
-
-
-from .header import MpvqcHeaderViewModel as MpvqcHeaderViewModel
-from .menu_bar import MpvqcMenuBarViewModel as MpvqcMenuBarViewModel
-from .toolbar import MpvqcToolBarViewModel as MpvqcToolBarViewModel
