+++ pyproject.toml
@@ -136,6 +136,7 @@ files = [
     "mpvqc/appearance/services/settings.py",
     "mpvqc/appearance/viewmodels/__init__.py",
     "mpvqc/appearance/viewmodels/dialog.py",
+    "mpvqc/appearance/viewmodels/fonts.py",
     "mpvqc/appearance/viewmodels/palette.py",
     "mpvqc/appearance/wiring.py",
     "mpvqc/application.py",
@@ -168,10 +169,6 @@ files = [
     "mpvqc/comments/viewmodels/selection.py",
     "mpvqc/comments/viewmodels/table_time_format.py",
     "mpvqc/comments/wiring.py",
-    "mpvqc/dialogs/__init__.py",
-    "mpvqc/dialogs/about.py",
-    "mpvqc/dialogs/edit_input.py",
-    "mpvqc/dialogs/edit_mpv.py",
     "mpvqc/exporting/__init__.py",
     "mpvqc/exporting/models/__init__.py",
     "mpvqc/exporting/models/templates.py",
@@ -232,9 +229,6 @@ files = [
     "mpvqc/injections.py",
     "mpvqc/jobs.py",
     "mpvqc/logging_utils.py",
-    "mpvqc/models/__init__.py",
-    "mpvqc/models/dependencies.py",
-    "mpvqc/models/shortcuts.py",
     "mpvqc/player/__init__.py",
     "mpvqc/player/services/__init__.py",
     "mpvqc/player/services/event_marshal.py",
@@ -270,26 +264,26 @@ files = [
     "mpvqc/shell/enums/message_box_kind.py",
     "mpvqc/shell/enums/time_display_mode.py",
     "mpvqc/shell/enums/window_title_format.py",
+    "mpvqc/shell/models/__init__.py",
+    "mpvqc/shell/models/dependencies.py",
+    "mpvqc/shell/models/shortcuts.py",
     "mpvqc/shell/services/__init__.py",
     "mpvqc/shell/services/quit.py",
     "mpvqc/shell/services/settings.py",
     "mpvqc/shell/services/vocabulary.py",
+    "mpvqc/shell/viewmodels/__init__.py",
+    "mpvqc/shell/viewmodels/about.py",
+    "mpvqc/shell/viewmodels/edit_input.py",
+    "mpvqc/shell/viewmodels/edit_mpv.py",
+    "mpvqc/shell/viewmodels/footer.py",
+    "mpvqc/shell/viewmodels/header.py",
+    "mpvqc/shell/viewmodels/menu_bar.py",
+    "mpvqc/shell/viewmodels/quit.py",
+    "mpvqc/shell/viewmodels/reset.py",
+    "mpvqc/shell/viewmodels/toolbar.py",
+    "mpvqc/shell/viewmodels/version_check.py",
     "mpvqc/shell/wiring.py",
     "mpvqc/startup.py",
-    "mpvqc/viewmodels/__init__.py",
-    "mpvqc/viewmodels/message_boxes/__init__.py",
-    "mpvqc/viewmodels/message_boxes/quit.py",
-    "mpvqc/viewmodels/message_boxes/request_relay.py",
-    "mpvqc/viewmodels/message_boxes/reset.py",
-    "mpvqc/viewmodels/message_boxes/version_check.py",
-    "mpvqc/viewmodels/utility/__init__.py",
-    "mpvqc/viewmodels/utility/fonts.py",
-    "mpvqc/viewmodels/views/__init__.py",
-    "mpvqc/viewmodels/views/footer.py",
-    "mpvqc/viewmodels/views/header/__init__.py",
-    "mpvqc/viewmodels/views/header/header.py",
-    "mpvqc/viewmodels/views/header/menu_bar.py",
-    "mpvqc/viewmodels/views/header/toolbar.py",
     "mpvqc/window/__init__.py",
     "mpvqc/window/services/__init__.py",
     "mpvqc/window/services/backend.py",
