+++ qt/qml/io/github/mpvqc/mpvQC/App/MpvqcOverlayController.qml
@@ -14,11 +14,25 @@ import io.github.mpvqc.mpvQC.Python
 Item {
     id: root
 
-    required property MpvqcMenuBarViewModel viewModel
+    required property MpvqcShellMenuBarViewModel viewModel
 
     signal focusWanted
-    signal closeAppRequested
-    signal resizeVideoRequested
+
+    function openDialog(kind: int): void {
+        _dialogLoader.openDialog(kind);
+    }
+
+    function openFileDialog(kind: int): void {
+        _fileDialogLoader.openFileDialog(kind);
+    }
+
+    function openCustomExportFileDialog(template: url): void {
+        _fileDialogLoader.openCustomExportFileDialog(template);
+    }
+
+    function openMessageBox(kind: int): void {
+        _messageBoxLoader.openMessageBox(kind);
+    }
 
     // Three loaders, never one: an export error opens a message box while the
     // file dialog tears down, and an UnfinishedPlan opens the wizard while
@@ -50,41 +64,19 @@ Item {
         onImportWizardRequested: viewModel => _dialogLoader.openImportWizardDialog(viewModel)
     }
 
-    MpvqcMessageBoxRequestRelayViewModel {
-        onExportErrorOccurred: (message, line) => _messageBoxLoader.openExportErrorMessageBox(message, line)
-
-        onQuitConfirmationNeeded: _messageBoxLoader.openMessageBox(MpvqcMessageBoxKind.MessageBoxKind.QUIT)
-    }
-
     Connections {
         target: root.viewModel
 
-        function onConfirmResetRequested(): void {
-            _messageBoxLoader.openMessageBox(MpvqcMessageBoxKind.MessageBoxKind.RESET);
-        }
-
         function onFileDialogRequested(kind: int): void {
-            _fileDialogLoader.openFileDialog(kind);
-        }
-
-        function onCustomExportRequested(template: url): void {
-            _fileDialogLoader.openCustomExportFileDialog(template);
-        }
-
-        function onCloseAppRequested(): void {
-            root.closeAppRequested();
-        }
-
-        function onResizeVideoRequested(): void {
-            root.resizeVideoRequested();
-        }
-
-        function onDialogRequested(kind: int): void {
-            _dialogLoader.openDialog(kind);
+            root.openFileDialog(kind);
         }
 
         function onMessageBoxRequested(kind: int): void {
-            _messageBoxLoader.openMessageBox(kind);
+            root.openMessageBox(kind);
+        }
+
+        function onExportErrorMessageBoxRequested(message: string, line: int): void {
+            _messageBoxLoader.openExportErrorMessageBox(message, line);
         }
     }
 }
