+++ test/shell/viewmodels/test_menu_bar.py
@@ -1,0 +1,172 @@
+# SPDX-FileCopyrightText: mpvQC developers
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+from pathlib import Path
+from unittest.mock import MagicMock
+
+import inject
+import pytest
+from PySide6.QtCore import Qt
+
+from mpvqc.comments.services import ResetService
+from mpvqc.exporting.services import ExportService
+from mpvqc.services import DesktopService, StateService
+from mpvqc.shell.enums import FileDialogKind, MessageBoxKind
+from mpvqc.shell.services import QuitService, ShellSettingsService
+from mpvqc.shell.viewmodels import MpvqcShellMenuBarViewModel
+
+MODULE = "mpvqc.shell.viewmodels.menu_bar"
+
+
+@pytest.fixture
+def reset_service_mock() -> MagicMock:
+    return MagicMock(spec_set=ResetService)
+
+
+@pytest.fixture
+def export_service_mock() -> MagicMock:
+    return MagicMock(spec_set=ExportService)
+
+
+@pytest.fixture
+def desktop_service_mock() -> MagicMock:
+    return MagicMock(spec_set=DesktopService)
+
+
+@pytest.fixture
+def quit_service() -> QuitService:
+    return QuitService()
+
+
+@pytest.fixture
+def view_model() -> MpvqcShellMenuBarViewModel:
+    # noinspection PyCallingNonCallable
+    return MpvqcShellMenuBarViewModel()
+
+
+@pytest.fixture(autouse=True)
+def configure_inject(
+    common_bindings_with,
+    reset_service_mock,
+    state_service,
+    shell_settings_service,
+    export_service_mock,
+    desktop_service_mock,
+    quit_service,
+):
+    def custom_bindings(binder: inject.Binder):
+        binder.bind(DesktopService, desktop_service_mock)
+        binder.bind(StateService, state_service)
+        binder.bind(ResetService, reset_service_mock)
+        binder.bind(ShellSettingsService, shell_settings_service)
+        binder.bind(ExportService, export_service_mock)
+        binder.bind(QuitService, quit_service)
+
+    common_bindings_with(custom_bindings)
+
+
+def test_request_reset_app_state(view_model, configure_state, reset_service_mock, make_spy):
+    spy = make_spy(view_model.messageBoxRequested)
+
+    configure_state(saved=True)
+    view_model.requestResetAppState()
+    reset_service_mock.reset.assert_called_once()
+    assert spy.count() == 0
+
+    reset_service_mock.reset.reset_mock()
+    configure_state(saved=False)
+    view_model.requestResetAppState()
+    assert spy.count() == 1
+    assert spy.at(0, 0) == MessageBoxKind.RESET
+    reset_service_mock.reset.assert_not_called()
+
+
+def test_request_quit_confirmation(view_model, quit_service, make_spy):
+    spy = make_spy(view_model.messageBoxRequested)
+
+    quit_service.confirmation_needed.emit()
+
+    assert spy.count() == 1
+    assert spy.at(0, 0) == MessageBoxKind.QUIT
+
+
+def test_request_export_error_message_box(common_bindings_with, shell_settings_service, quit_service, make_spy):
+    export_service = ExportService()
+
+    def custom_bindings(binder: inject.Binder):
+        binder.bind(ShellSettingsService, shell_settings_service)
+        binder.bind(ExportService, export_service)
+        binder.bind(QuitService, quit_service)
+
+    common_bindings_with(custom_bindings)
+    view_model = MpvqcShellMenuBarViewModel()
+    spy = make_spy(view_model.exportErrorMessageBoxRequested)
+
+    export_service.export_error_occurred.emit("message", 42)
+
+    assert spy.count() == 1
+    assert spy.at(0, 0) == "message"
+    assert spy.at(0, 1) == 42
+
+
+def test_save(view_model, make_spy, configure_state, export_service_mock):
+    spy = make_spy(view_model.fileDialogRequested)
+
+    configure_state(document=None)
+    view_model.requestSaveQcDocument()
+    assert spy.count() == 1
+    assert spy.at(0, 0) == FileDialogKind.SAVE_DOCUMENT
+    export_service_mock.save.assert_not_called()
+
+    path = Path() / "test_document.txt"
+    configure_state(document=path)
+    view_model.requestSaveQcDocument()
+    assert spy.count() == 1
+    assert export_service_mock.save.call_count == 1
+    export_service_mock.save.assert_called_with(path)
+
+
+def test_open_app_data_folder(view_model, desktop_service_mock):
+    view_model.openAppDataFolder()
+
+    desktop_service_mock.open_app_data_folder.assert_called_once_with()
+
+
+def test_configure_layout_orientation(view_model, shell_settings_service, make_spy):
+    spy = make_spy(view_model.layoutOrientationChanged)
+    horizontal = Qt.Orientation.Horizontal.value
+
+    view_model.configureLayoutOrientation(horizontal)
+
+    assert shell_settings_service.layout_orientation == horizontal
+    assert view_model.layoutOrientation == horizontal
+    assert spy.count() == 1
+    assert spy.at(0, 0) == horizontal
+
+
+@pytest.mark.parametrize(
+    ("debug_env", "offers_update_check", "expected"),
+    [
+        (None, False, False),
+        (None, True, True),
+        ("1", False, True),
+        ("1", True, True),
+    ],
+)
+def test_is_update_menu_visible(
+    view_model,
+    make_build_info,
+    monkeypatch,
+    debug_env,
+    offers_update_check,
+    expected,
+):
+    if debug_env is None:
+        monkeypatch.delenv("MPVQC_DEBUG", raising=False)
+    else:
+        monkeypatch.setenv("MPVQC_DEBUG", debug_env)
+    info = make_build_info(offers_update_check=offers_update_check)
+    monkeypatch.setattr(f"{MODULE}.get_build_info", lambda: info)
+
+    assert view_model.isUpdateMenuVisible is expected
